+++ src/test/kotlin/cz/loplex/reflog/GitReflogTabTest.kt
@@ -26,6 +26,7 @@ class GitReflogTabTest : BasePlatformTestCase() {
 
         assertNotNull(actionManager.getAction("GitReflog.Toolbar"))
         assertNotNull(actionManager.getAction("GitReflog.ContextMenu"))
+        assertNotNull(actionManager.getAction("GitReflog.LoadMore"))
         assertNotNull(actionManager.getAction("GitReflog.ShowDiff"))
         assertNotNull(actionManager.getAction("GitReflog.CheckoutRevision"))
         assertNotNull(actionManager.getAction("GitReflog.NewBranch"))